test(pg_cron): split the grammar rule from its agreement with the extension
The rule tables stay in validators/, pure Python, named for the validator they
exercise. What the installed extension answers moves out to its own module,
named for the question it asks; it needed the central catalog, which nothing
else under validators/ does.

Its our-side rejection test went with the split: the rule module already
asserts those, and this module now only reports what pg_cron says.
